v.0.5.0

+++ view/convenio/lista.php
@@ -68,9 +68,12 @@
             echo "<td>$al->duracion</td>";
             echo "<td>$al->detalles</td>";
             echo "<td><a href='/convenio/show/$al->id'>Ver</a></td>";
+            if(Login::get()->rol =="cordinador"){
+                
+         
             echo "<td><a href='/convenio/edit/$al->id'>Actualizar</a></td>";
             echo "<td><a href='/convenio/delete/$al->id'>Eliminar</a></td>";
-           
+            }
          
             echo "</td>";
             echo "</tr>";
